test-factorial environment

+++ 2025-05-14/factorial.h
@@ -1,2 +1,5 @@
 #pragma once
-int factorial(int n);
+
+#include<iostream>
+
+long factorial(long n);
